+++ HISTORY.rst
@@ -18,6 +18,7 @@ Unreleased
 
   * Fix a bug (set the right name) with the additional line (product with an additional product) in a purchase order
   * Remove lines with a quantity == 0 in the purchase order report.
+* ALCYN-1417: Fix ESB connector following addition of customer hierarchy
 
 **Bugfixes**