+++ HISTORY.rst
@@ -27,6 +27,8 @@ Unreleased
 
 **Build**
 
+* Upgrade project to 10.0-2.7.0
+
 **Documentation**